Only CDD providers can add CDD claims
