+++ pallets/runtime/src/test/mod.rs
@@ -8,6 +8,7 @@ mod asset_test;
 mod balances_test;
 mod bridge;
 mod committee_test;
+mod general_tm_test;
 mod group_test;
 mod identity_test;
 mod multisig;
